+++ src/main/java/cn/elytra/mod/nomi_horizons/mixins/gt/ElevatorInCleanroom_Mixin.java
@@ -1,0 +1,24 @@
+package cn.elytra.mod.nomi_horizons.mixins.gt;
+
+import cn.elytra.mod.nomi_horizons.NomiHorizons;
+import gregtech.api.pattern.TraceabilityPredicate;
+import gregtech.common.metatileentities.multi.electric.MetaTileEntityCleanroom;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Unique;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Redirect;
+
+@Mixin(value = MetaTileEntityCleanroom.class, remap = false)
+public class ElevatorInCleanroom_Mixin {
+
+    @Unique
+    private static TraceabilityPredicate nh$elevatorPredicate() {
+        return new TraceabilityPredicate(blockWorldState -> NomiHorizons.API.getElevatorPredicate().test(blockWorldState.getBlockState())).setMaxGlobalLimited(1);
+    }
+
+    @Redirect(method = "createStructurePattern", at = @At(value = "INVOKE", target = "Lgregtech/api/pattern/TraceabilityPredicate;or(Lgregtech/api/pattern/TraceabilityPredicate;)Lgregtech/api/pattern/TraceabilityPredicate;", ordinal = 2))
+    private TraceabilityPredicate nh$injectElevatorPredicate(TraceabilityPredicate instance, TraceabilityPredicate traceabilityPredicate) {
+        return instance.or(nh$elevatorPredicate()).or(traceabilityPredicate);
+    }
+
+}